Fixes error on dotty remove command
When you attempt to remove a repo using `dotty remove` it has an error
complaining that the variable "name" can not be modified because it is
frozen.

When looking at the code it seems that the find_repo! method attepts to
use .downcase! on the name variable that is passed in. Assuming that the
passed argument is readonly this would cause such an error. Looking
elsewhere in the code it seems that find_repo! is offten called as
`find_repo!(name.downcase)` and then it is mutated again in that method
a second time.

By assigning a temp variable the value we avoid the frozen error
message.